+++ src/dimension.cpp
@@ -69,7 +69,7 @@ veci2 dimension::searchPortal(cveci2& positionNear)
 {
 	cfp netherPortalSearchRange = identifier == dimensionID::nether ? netherPortalSearchRangeNether : netherPortalSearchRangeOverWorld;
 	fp minimumDistanceSquared = INFINITY;
-	int portalIndex = 0;
+	size_t portalIndex = 0;
 	for (size_t i = 0; i < portalPositions.size(); i++)
 	{
 		veci2 location = portalPositions[i];
